+++ backend/apps/batch_printing/apps.py
@@ -5,3 +5,6 @@ class BatchPrintingConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "apps.batch_printing"
     verbose_name = "批量打印"
+
+    def ready(self) -> None:
+        from . import signals  # noqa: F401
